Move third-party harvester protocol tests
Moved to own directory in order to enable `install_timelord = True`
