+++ tests/farmer_harvester/third_party_harvesters/config.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+install_timelord = True
